+++ both/models/User.jsx
@@ -1,111 +1,12 @@
 Users = Meteor.users;
 
-// adapted from https://github.com/aldeed/meteor-collection2 example
-Schema.UserProfile = {
-    name: {
-      type: String,
-      label: "Display name"
-    },
-    email: {
-      type: String,
-      label: "Email",
-      //regEx: SimpleSchema.RegEx.Email
-    },
-    phone: {
-      type: String,
-      label: "Phone",
-      optional:true
-    },
-    thumb: {
-      type: String,
-      label: "Thumbnail",
-      optional:true
-    },
-    firstName: {
-        type: String,
-        label: "First name",
-        optional: true
-    },
-    lastName: {
-        type: String,
-        label: "Last Name",
-        optional: true
-    },
-    dob: {
-        type: Date,
-        label: "DOB",
-        optional: true
-    },
-    website: {
-        type: String,
-        //regEx: SimpleSchema.RegEx.Url,
-        optional: true
-    },
-    bio: {
-        type: String,
-        label: "About me",
-        optional: true
-    }
+var template = {
+  profile:{
+    name:"",
+    firstName:"",
+    lastName:""
+  }
 }
-
-Schema.User = {
-    username: {
-        type: String,
-        optional: true
-    },
-    emails: {
-        type: Array,
-        optional: false
-    },
-    "emails.$": {
-        type: Object
-    },
-    "emails.$.address": {
-        type: String,
-        //regEx: SimpleSchema.RegEx.Email
-    },
-    "emails.$.verified": {
-        type: Boolean
-    },
-    createdAt: {
-        type: Date,
-        autoValue:function(){
-          return moment().toDate();
-        }
-    },
-    profile: {
-        type: Schema.UserProfile,
-        optional: true
-    },
-    // Make sure this services field is in your schema if you're using any of the accounts packages
-    services: {
-        type: Object,
-        optional: true,
-        blackbox: true
-    },
-    // Add `roles` to your schema if you use the meteor-roles package.
-    // Option 1: Object type
-    // If you specify that type as Object, you must also specify the
-    // `Roles.GLOBAL_GROUP` group whenever you add a user to a role.
-    // Example:
-    // Roles.addUsersToRoles(userId, ["admin"], Roles.GLOBAL_GROUP);
-    // You can't mix and match adding with and without a group since
-    // you will fail validation in some cases.
-    roles: {
-        type: Object,
-        optional: true,
-        blackbox: true
-    },
-    // Option 2: [String] type
-    // If you are sure you will never need to use role groups, then
-    // you can specify [String] as the type
-    roles: {
-        type: [String],
-        optional: true
-    }
-}
-
-//Meteor.users.attachSchema(Schema.User);
 
 Meteor.methods({
   "User.save": function(item) {
@@ -127,6 +28,9 @@ Meteor.methods({
       user.password = password;
     }
     return Accounts.createUser(user);
+  },
+  "User.getTemplate":function(item) {
+    return _.extend({},template,item);
   }
 })