+++ CHANGELOG.md
@@ -2,9 +2,15 @@ CHANGELOG.md
 ============
 
 **v0.3.5a**
+
 * Clicking away from a dropdown closes it
-* WO: Facilities appearing in drop down for new work order should are filtered by selected team
-* WO: Available areas are filtered by selected facility
+* New feature - improved facilities selector
+** WO: Facilities appearing in drop down for new work order should are filtered by selected team
+** WO: Available areas are filtered by selected facility
+* New feature - can add and remove contacts
+** Facility settings: Can add new contacts and tenants to facilities
+** Facility settings: Can remove contacts and tenanats from facilities
+** Team: Pushing (+) button brings modal to invite new or add existing user
 
 **v0.3.4a**